Release 2.2.0 * Changes license to MPL 2.0

+++ src/Tilengine.c
@@ -1,21 +1,12 @@
 /*
 * Tilengine - The 2D retro graphics engine with raster effects
-* Copyright (C) 2015-2018 Marc Palacios Domenech <mailto:megamarc@hotmail.com>
+* Copyright (C) 2015-2019 Marc Palacios Domenech <mailto:megamarc@hotmail.com>
 * All rights reserved
 *
-* This library is free software; you can redistribute it and/or
-* modify it under the terms of the GNU Lesser General Public
-* License as published by the Free Software Foundation; either
-* version 2 of the License, or (at your option) any later version.
-*
-* This library is distributed in the hope that it will be useful,
-* but WITHOUT ANY WARRANTY; without even the implied warranty of
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-* Library General Public License for more details.
-*
-* You should have received a copy of the GNU Library General Public
-* License along with this library. If not, see <http://www.gnu.org/licenses/>.
-*/
+* This Source Code Form is subject to the terms of the Mozilla Public
+* License, v. 2.0. If a copy of the MPL was not distributed with this
+* file, You can obtain one at http://mozilla.org/MPL/2.0/.
+* */
 
 #include <string.h>
 #include <stdlib.h>
@@ -64,7 +55,7 @@ static TLN_Engine create_context(int hres, int vres, int bpp, int numlayers, int
  */
 TLN_Engine TLN_Init (int hres, int vres, int numlayers, int numsprites, int numanimations)
 {
-	printf("Tilengine v%d.%d.%d %d-bit built %s %s\n", TILENGINE_VER_MAJ, TILENGINE_VER_MIN, TILENGINE_VER_REV, sizeof(UINTPTR_MAX) << 3, __DATE__, __TIME__);
+	printf("Tilengine v%d.%d.%d %d-bit built %s %s\n", TILENGINE_VER_MAJ, TILENGINE_VER_MIN, TILENGINE_VER_REV, (int)(sizeof(UINTPTR_MAX) << 3), __DATE__, __TIME__);
 	return create_context (hres, vres, 32, numlayers, numsprites, numanimations);
 }
